refactor : modalsProvider, useModal ts로 전환
