+++ src/hooks/useModals.tsx
@@ -4,11 +4,11 @@ import { ModalsDispatchContext } from '@/contexts';
 export default function useModals() {
   const { open, close } = useContext(ModalsDispatchContext);
 
-  const openModal = (Component, props) => {
+  const openModal = (Component: React.ReactElement, props: any[]) => {
     open(Component, props);
   };
 
-  const closeModal = Component => {
+  const closeModal = (Component: React.ReactElement) => {
     close(Component);
   };
