+++ src/types/index.ts
@@ -1,2 +1,3 @@
 export type { UserType } from './user';
 export type { WaffleCardType } from './waffleCard';
+export type { ModalsStateType } from './modals';
