+++ src/types/modals.ts
@@ -1,0 +1,4 @@
+export type ModalsStateType = {
+  Component: React.ReactElement;
+  props: any[];
+};
